requirements.txt
elasticsearch>=8.0.0
python-dotenv==1.0.0
httpx==0.25.0
pydantic==2.5.0
pydantic-settings==2.1.0
flask==3.0.0
